feat: instantly refresh the sender's scores
Players like to run these commands as soon as they submit a score, but
since the tracker only tracks every minute it would have a slight delay.

The likelihood that 2 players play the same map and submitted the score
within a minute is pretty low, so we don't have to refresh the whole
thing. I doubt people would really complain about the short desync
anyway.